Refactor to get POEXTENDED accessible from other modules
