+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/ios/ios_generate_po_file_from_metadata.rb
@@ -1,31 +1,9 @@
+require_relative '../../helper/poextended.rb'
 require 'fastlane/action'
 require 'gettext/po'
 
 module Fastlane
   module Actions
-
-    class POMine < GetText::PO
-
-      def sort_by_msgctxt(entries)
-        entries.sort_by do |msgid_entry|
-          # msgid_entry = [[msgctxt, msgid], POEntry]
-          msgid_entry[1][0]
-        end
-      end
-
-      def sort(entries)
-        case @order
-        when :reference, :references # :references is deprecated.
-          sort_by_reference(entries)
-        when :msgid
-          sort_by_msgid(entries)
-        when :msgctxt
-          sort_by_msgctxt(entries)
-        else
-          entries.to_a
-        end
-      end
-    end
     class IosGeneratePoFileFromMetadataAction < Action
       # TODO: release_notes_previous.txt must be generated by the code_freeze lanes
 
@@ -73,7 +51,7 @@ module Fastlane
           standard_files.append(key) unless SPECIAL_KEYS.include? File.basename(key, '.txt')
         end
         # Let the helper handle standard files
-        @po = POMine.new(:msgctxt)
+        @po = POEXTENDED.new(:msgctxt)
         @po = Fastlane::Helper::GeneratePoFileMetadataHelper.add_standard_files_to_po(prefix, files: standard_files, keys_to_comment_hash: KEYS_TO_COMMENT_HASH, po_obj: @po)
 
         other_sources_files = []
