+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/poextended.rb
@@ -1,0 +1,23 @@
+require 'gettext/po'
+
+class POEXTENDED < GetText::PO
+  def sort_by_msgctxt(entries)
+    entries.sort_by do |msgid_entry|
+      # msgid_entry = [[msgctxt, msgid], POEntry]
+      msgid_entry[1][0]
+    end
+  end
+
+  def sort(entries)
+    case @order
+    when :reference, :references # :references is deprecated.
+      sort_by_reference(entries)
+    when :msgid
+      sort_by_msgid(entries)
+    when :msgctxt
+      sort_by_msgctxt(entries)
+    else
+      entries.to_a
+    end
+  end
+end
